% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hazards_to_life_expectancy.R
\name{convert_hazards_to_e65}
\alias{convert_hazards_to_e65}
\title{Translate to e65}
\usage{
convert_hazards_to_e65(df)
}
\arguments{
\item{lower}{the bound function}

\item{upper}{the initialization function}

\item{b}{the hazard ratio}
}
\value{
None
}
\description{
Translate to e65
}
