% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hazards_to_ex.R
\name{convert_hazards_to_ex}
\alias{convert_hazards_to_ex}
\title{Convert hazard ratios to life expectancy}
\usage{
convert_hazards_to_ex(
  df,
  age = 65,
  upper_age = 120,
  M = 80,
  b = 0.075,
  use_model_estimates = F
)
}
\arguments{
\item{df}{dataframe of results given by gompertz_mle() function}

\item{age}{age at which to calculate remaining life expectancy}

\item{upper_age}{maximal age to use in life table calculation}

\item{M}{Gompertz parameter modal age at death}

\item{b}{Gompertz mortality slope parameter}

\item{use_model_estimates}{use estimates of the Gompertz Parameters from the model, rather than defaults}
}
\value{
A dataframe of hazards ratios and corresponding e(x) estimates and confidence intervals
}
\description{
Convert hazard ratios to differences in remaining life expectancy
at a given age (defaults to age 65)
}
