% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hazards_to_life_expectanct.R
\name{convert_hazards_to_life_expectancy}
\alias{convert_hazards_to_life_expectancy}
\title{Translate to e65}
\usage{
convert_hazards_to_life_expectancy(
  df,
  lower = 65,
  upper = 120,
  M = 80,
  beta = 0.1
)
}
\arguments{
\item{lower}{the bound function}

\item{upper}{the initialization function}

\item{b}{the hazard ratio}
}
\value{
None
}
\description{
Translate to e65
}
