% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot_hazard.R
\name{diagnostic_plot_hazard}
\alias{diagnostic_plot_hazard}
\title{Run diagnostic plots - plot hazards}
\usage{
diagnostic_plot_hazard(
  data,
  object,
  death_var = "death_age",
  covar,
  xlim = c(65, 110)
)
}
\arguments{
\item{data}{data.frame use for gompertz_mle}

\item{object}{gompertz_mle object}

\item{death_var}{death age variable}

\item{covar}{covariate of interest}

\item{xlim}{x-limits for figure}
}
\value{
a ggplot object
}
\description{
Compare empirical and model-based estimated hazard ratios within a cohort. Only
works with a single discrete covariate.
}
\examples{

# Create a single-cohort data set
numident_c1920 <- numident_demo \%>\% dplyr::filter(byear == 1920) \%>\%
dplyr::mutate(finished_hs = as.factor(educ_yrs >= 12))

# Run gompertz_mle()
gradient <- gompertztrunc::gompertz_mle(formula = death_age ~ finished_hs,
left_trunc = 1988, right_trunc = 2005, data = numident_c1920)

# Create diagnostic hazards plot using model outcome
gompertztrunc::diagnostic_plot_hazard(object = gradient, data = numident_c1920,
covar = "finished_hs", xlim = c(60, 95))
}
