% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertztrunc_functions.R,
%   R/optimizer_function_v2.R
\name{get.negLL}
\alias{get.negLL}
\title{Gompertz mle function}
\usage{
get.negLL(par, y, X, y.left, y.right, wt = 1)

get.negLL(par, y, X, y.left, y.right, wt = 1)
}
\arguments{
\item{fml}{the estimation formula}

\item{A}{data matrix with covariates y, u, l, and covariates, including cohor}
}
\value{
None

None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
