% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.par.start.R, R/get_starting_param.R
\name{get.par.start}
\alias{get.par.start}
\title{Get starting values for parameters}
\usage{
get.par.start(form, da)

get.par.start(form, da)
}
\arguments{
\item{fml}{the estimation formula}

\item{data}{data matrix with covariates y, u, l, and covariates, including cohort}

\item{p}{the initialization function}
}
\value{
Named vector of initial parameter estimates

None
}
\description{
Uses linear modeling to compute initial values for MLE optimizer
}
