% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_ratio_to_le.R
\name{hazard_ratio_to_le}
\alias{hazard_ratio_to_le}
\title{Translate a single hazard ratio to life expectancy}
\usage{
hazard_ratio_to_le(lower, upper, hr, M = 80, b = 0.1)
}
\arguments{
\item{lower}{lower age bound}

\item{upper}{upper age bound}

\item{hr}{hazard ratio}

\item{M}{Gompertz modal age at death parameter}

\item{b}{Gompertz mortality slope parameter}
}
\value{
hazard ratio converted to effect on e65
}
\description{
Translate a single hazard ratio to life expectancy
}
