% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_plot.R
\name{mle_plot}
\alias{mle_plot}
\title{Run diagnostic plots}
\usage{
mle_plot(data, object, var = hs, death_range = c(65, 110))
}
\arguments{
\item{data}{data from model plot}

\item{object}{gompertz_mle object}

\item{var}{variable of interest}
}
\value{
a ggplot object
}
\description{
Empirical and modeled distribution of ages of death.
}
