% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_mle_optim_function.R
\name{mll.gomp.multi.cohort.cov}
\alias{mll.gomp.multi.cohort.cov}
\title{Gompertz mle function}
\usage{
mll.gomp.multi.cohort.cov(p, A, predictors = predictors)
}
\arguments{
\item{p}{the initialization function}

\item{A}{data matrix with covariates y, u, l, and covariates, including cohort}
}
\value{
None
}
\description{
Estimates separate alpha values for each cohort
}
\keyword{internal}
