% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negLL_function_baseline.R
\name{negLL_function_baseline}
\alias{negLL_function_baseline}
\title{Gompertz Negative Log Likelihood Function}
\usage{
negLL_function_baseline(par, y, X, y.left, y.right, wt)
}
\arguments{
\item{par}{a vector of parameter estimates}

\item{y}{a vector of death ages}

\item{X}{a model matrix}

\item{y.left}{left truncation age}

\item{y.right}{right trunction age}

\item{wt}{weight}
}
\value{
The negative log likelihood of parameter estimates given observed data
}
\description{
Computes negative log likelihood for optimizer
}
