% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs, rhs}{specify what lhs and rhs are}

\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling `rhs(lhs)`.
}
\description{
Put description here

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\examples{
# some examples if you want to highlight the usage in the package
}
\keyword{internal}
