% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated mortality data set}
\format{
A data frame with 6732 rows and 6 variables:
\describe{
\item{aod}{Age at death, in integer years}
\item{byear}{Calendar year of birth}
\item{dyear}{Calendar year of death}
\item{temp}{Temperature}
\item{sex}{Sex (0 = male, 1 = female)}
\item{isSouth}{Live in south (0 = FALSE, 1 = TRUE)}

}
}
\usage{
sim_data
}
\description{
A data set containing simulated age at death and covariates according to
a truncated Gompertz distribution with proportional hazards
}
\keyword{datasets}
