% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/josh_gompertz_functions.R
\name{tgm_simu}
\alias{tgm_simu}
\title{Gompertz mle functions from Josh}
\usage{
tgm_simu(
  n,
  form,
  coefs,
  dummy = NULL,
  sigma = NULL,
  seed = NULL,
  a0 = 10^-4,
  beta = 1/10,
  verbose = FALSE
)
}
\value{
None
}
\description{
Gompertz mle functions from Josh
}
