% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgm_summary.R
\name{tgm_summary}
\alias{tgm_summary}
\title{Gompertz mle function}
\usage{
tgm_summary(fit, true.coef.values = NULL, a0 = 10^-4, beta = 1/10)
}
\arguments{
\item{fml}{the estimation formula}

\item{A}{data matrix with covariates y, u, l, and covariates, including cohor}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
