% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunmd_demo.R
\docType{data}
\name{bunmd_demo}
\alias{bunmd_demo}
\title{Demo BUNMD Data Set}
\format{
A data frame with 81,002 rows and 6 variables:
\describe{
  \item{ssn}{Social Security number}
  \item{bpl_string}{Country of birth}
  \item{death_age}{Age at death (integer years)}
  \item{byear}{Calender year of birth}
  \item{dyear}{Calender year of death}
  \item{age_first_application}{Age at first Social Security application}

}
}
\usage{
bunmd_demo
}
\description{
A data set containing a sample of the CenSoc Berkeley Unified Numident Mortality Database (BUNMD) file,
including age at death and select covariates.
}
\keyword{datasets}
