% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoc_example.R
\docType{data}
\name{censoc_example}
\alias{censoc_example}
\title{Example CenSoc Data Set}
\format{
A data frame with 100000 rows and 6 variables:
\describe{
  \item{id}{Individual-level identifier}
  \item{byear}{Calender year of birth}
  \item{death_age}{Age at death (integer years)}
  \item{educ_yrs}{Years of education completed}
  \item{urban}{Residence in urban area in 1940 (0 = FALSE, 1 = TRUE)}
  \item{ownhome}{Homeowner (0 = FALSE, 1 = TRUE)}
}
}
\usage{
censoc_example
}
\description{
A data set containing age at death and select covariates
for sample for men born in 1910.
}
\keyword{datasets}
