% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hazard_ratios_to_le.R
\name{convert_hazard_ratio_to_le}
\alias{convert_hazard_ratio_to_le}
\title{Translate to e65}
\usage{
convert_hazard_ratio_to_le(lower, upper, hr)
}
\arguments{
\item{lower}{the bound function}

\item{upper}{the initialization function}

\item{b}{the hazard ratio}
}
\value{
None
}
\description{
Translate to e65
}
