% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hazards_to_e65.R
\name{convert_hazards_to_e65}
\alias{convert_hazards_to_e65}
\title{Convert hazard ratios to life expectancy at age 65}
\usage{
convert_hazards_to_e65(
  df,
  lower = 65,
  upper = 120,
  M = 80,
  b = 0.075,
  use_model_estimates = F
)
}
\arguments{
\item{df}{dataframe of results given by gompertz_mle() function}

\item{lower}{lower age bound}

\item{upper}{upper age bound}

\item{M}{Gompertz parameter modal age at death}

\item{b}{Gompertz mortality slope parameter}

\item{use_model_estimates}{use estimates of the Gompertz Parameters from the model, rather than defaults}
}
\value{
A dataframe of hazards ratios and corresponding e65 estimates and confidence intervals
}
\description{
Convert hazard ratios to life expectancy at age 65
}
