% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot.R
\name{diagnostic_plot}
\alias{diagnostic_plot}
\title{Run diagnostic plots}
\usage{
diagnostic_plot(
  data,
  object,
  covar,
  xlim = c(65, 110),
  death_var = "death_age"
)
}
\arguments{
\item{data}{data.frame use for gompertz_mle}

\item{object}{gompertz_mle object}

\item{covar}{covariate of interest}

\item{xlim}{x-limits for figure}

\item{death_var}{death age variable}
}
\value{
a ggplot object
}
\description{
Compare empirical and model-based distribution of ages of death within a cohort. Only
works with a single discrete covariate.
}
\examples{

# Diagnostic plot to assess model fit. This diagnostic plot only works for one birth cohort
# and discrete covariates.
gompertztrunc::diagnostic_plot(object = gradient, data = sim_data_trunc, covar = "sex",
death_range = c(65, 85))

}
