% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot_hazard.R
\name{diagnostic_plot_hazard}
\alias{diagnostic_plot_hazard}
\title{Run diagnostic plots - plot hazards}
\usage{
diagnostic_plot_hazard(
  data,
  object,
  death_var = "death_age",
  covar = hs,
  xlim = c(65, 110)
)
}
\arguments{
\item{data}{data.frame use for gompertz_mle}

\item{object}{gompertz_mle object}

\item{death_var}{death age variable}

\item{covar}{covariate of interest}

\item{xlim}{x-limits for figure}
}
\value{
a ggplot object
}
\description{
Compare empirical and model-based estimated hazard ratios within a cohort. Only
works with a single discrete covariate.
}
\examples{
# Diagnostic hazard plot to assess model fit. This diagnostic plot only works for one birth cohort
# and discrete covariates.
gompertztrunc::diagnostic_plot_hazard(object = gradient, data = sim_data_trunc, covar = "sex",
 death_range = c(65, 85))

}
