% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.par.start.R, R/get_starting_param.R
\name{get.par.start}
\alias{get.par.start}
\title{Gompertz mle function}
\usage{
get.par.start(form, da)

get.par.start(form, da)
}
\arguments{
\item{fml}{the estimation formula}

\item{A}{data matrix with covariates y, u, l, and covariates, including cohor}

\item{p}{the initialization function}
}
\value{
None

None
}
\description{
Gompertz mle function

Gompertz mle function
}
