% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.par.start.R, R/get_starting_param.R
\name{get.par.start}
\alias{get.par.start}
\title{Get starting values for parameters}
\usage{
get.par.start(formula, da)

get.par.start(formula, da)
}
\arguments{
\item{formula}{the estimation formula}

\item{data}{data frame including death age, cohort, truncation bounds, and covariates}
}
\value{
Named vector of initial parameter estimates

Named vector of initial parameter values for optimizer
}
\description{
Uses linear modeling to compute initial values for MLE optimizer

Uses linear regression to set initial values for parameters to be estimated
}
