% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_param.R
\name{get.par.start}
\alias{get.par.start}
\title{Gompertz mle function}
\usage{
get.par.start(form, da)
}
\arguments{
\item{p}{the initialization function}
}
\value{
None
}
\description{
Gompertz mle function
}
