% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_mle.R
\name{gompertz_mle}
\alias{gompertz_mle}
\title{Gompertz mle function}
\usage{
gompertz_mle(fml, upper = 2005, lower = 1975, data, byear = byear)
}
\arguments{
\item{fml}{the estimation formula}

\item{A}{data matrix with covariates y, u, l, and covariates, including cohor}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
