% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_mle.R
\name{gompertz_mle}
\alias{gompertz_mle}
\title{Gompertz mle function}
\usage{
gompertz_mle(
  fml,
  right_trunc = 2005,
  left_trunc = 1975,
  data,
  byear = byear,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{fml}{the estimation formula}

\item{A}{data matrix with covariates y, u, l, and covariates, including cohor}
}
\value{
fit
}
\description{
Gompertz mle function
}
