% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_mle.R
\name{gompertz_mle}
\alias{gompertz_mle}
\title{Gompertz MLE function}
\usage{
gompertz_mle(
  formula,
  left_trunc = 1975,
  right_trunc = 2005,
  data,
  byear = byear,
  dyear = dyear,
  lower_age_bound = NULL,
  upper_age_bound = NULL,
  weights = NULL,
  death_age_data_type = "auto",
  maxiter = 10000
)
}
\arguments{
\item{formula}{the estimation formula}

\item{left_trunc}{left truncation year}

\item{right_trunc}{right truncation year}

\item{data}{a data frame containing variables in the model}

\item{byear}{name of birth year variable in \code{data}}

\item{dyear}{name of death year variablein \code{data}}

\item{lower_age_bound}{lowest age at death (optional)}

\item{upper_age_bound}{highest age at death (optional)}

\item{weights}{an optional vector of individual weights}

\item{death_age_data_type}{option for handling of continuous and discrete death age variable (not yet implemented)}

\item{maxiter}{maximum number of iterations for optimizer}
}
\value{
Returns a named list consisting of the following components
(See \code{stats::\link[stats:optim]{optim}} for additional details):
\describe{
  \item{\code{starting_values}}{list of starting values of parameters}
  \item{\code{optim_fit}}{A list consisting of:
  \describe{
    \item{\code{par}}{best estimation of parameter values}
    \item{\code{value}}{log likelihood}
    \item{\code{counts}}{number of calls to funcation and gradient}
    \item{\code{convergence}}{returns 0 if the model converged, for other values see \code{stats::\link[stats:optim]{optim}} }
    \item{\code{message}}{any other information returned by optimizer}
    \item{\code{hessian}}{Hessian matrix}
    }
  }
  \item{\code{results}}{a table of estimates and upper/lower bounds}
}
}
\description{
Fits a truncated Gompertz distribution with proportional hazards
to mortality data using maximum likelihood estimation.
}
\examples{
#model hazards as function of birthplace using bunmd_demo file
gompertz_mle(formula = death_age ~ bpl_string, left_trunc = 1988, right_trunc = 2005, data = bunmd_demo)

}
