% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertztrunc_simu.R
\name{gompertztrunc_simu}
\alias{gompertztrunc_simu}
\title{Simulate Gompertz death distribution}
\usage{
gompertztrunc_simu(
  n,
  form,
  coefs,
  dummy = NULL,
  sigma = NULL,
  seed = NULL,
  a0 = 10^-4,
  beta = 1/10,
  verbose = FALSE
)
}
\arguments{
\item{n}{sample size}

\item{form}{estimation formula}

\item{coefs}{named vactors of coefficients and corresponding true values}

\item{dummy}{vector flags for each coefficient}

\item{sigma}{standard deviation for each variable}

\item{seed}{random seed to duplicate data}

\item{a0}{Gompertz alpha parameter}

\item{beta}{Gompertz beta parameter}

\item{verbose}{print internal check if true}
}
\value{
data frame of simulated death ages and covariate values
}
\description{
Simulate Gompertz death distribution
}
\examples{
gompertztrunc_simu(n=1000, form = death_age ~ sex + edu,
coefs = c('sex'=-0.8, 'ambient_temp'=0.3), dummy=c(T,F))

}
