% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_ratio_to_le.R
\name{hazard_ratio_to_le}
\alias{hazard_ratio_to_le}
\title{Translate a single hazard ratio to remaining life expectancy}
\usage{
hazard_ratio_to_le(lower, upper, hr, M = 80, b = 0.1)
}
\arguments{
\item{lower}{age at which to compute change in remaining life expectancy}

\item{upper}{upper age bound for life table calculations}

\item{hr}{hazard ratio}

\item{M}{Gompertz modal age at death parameter}

\item{b}{Gompertz mortality slope parameter}
}
\value{
hazard ratio converted to effect on life expectancy
}
\description{
Translate a single hazard ratio to effect on remaining life
expectancy at a specified age, using a Gompertz mortality schedule
as the baseline
}
