% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negLL_function.R
\name{negLL_function}
\alias{negLL_function}
\title{Gompertz mle function}
\usage{
negLL_function(par, y, X, y.left, y.right, wt = 1)
}
\arguments{
\item{fml}{the estimation formula}

\item{A}{data matrix witdevh covariates y, u, l, and covariates, including cohor}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
