% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tig_simu.R
\name{tgm_simu}
\alias{tgm_simu}
\title{Gompertz mle function}
\usage{
tgm_simu(
  n,
  form,
  coefs,
  dummy = NULL,
  sigma = NULL,
  seed = NULL,
  a0 = 10^-4,
  beta = 1/10,
  verbose = FALSE
)
}
\arguments{
\item{fml}{the estimation formula}

\item{A}{data matrix with covariates y, u, l, and covariates, including cohor}
}
\value{
None
}
\description{
Gompertz mle function
}
